+++ glean-core/ios/Glean/Metrics/TimespanMetric.swift
@@ -85,6 +85,22 @@ public class TimespanMetricType {
         }
     }
 
+    /// Convenience method to simplify measuring a function or block of code
+    ///
+    /// - parameters:
+    ///     * funcToMeasure: Accepts a function or closure  to measure that can return a value
+    public func measure<U>(funcToMeasure: () -> U) -> U {
+        start()
+        // Putting `stop` in a `defer` block guarantees it will execute at the end
+        // of the scope, after the return value is pushed onto the stack.
+        // Reference: https://docs.swift.org/swift-book/LanguageGuide/ErrorHandling.html
+        // under the "Specifying Cleanup Actions" section.
+        defer {
+            stop()
+        }
+        return funcToMeasure()
+    }
+
     /// Explicitly set the timespan value, in nanoseconds.
     ///
     /// This API should only be used if your library or application requires recording