+++ glean-core/python/glean/metrics/__init__.py
@@ -3,17 +3,30 @@
 # file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
 
+"""
+This module contains all of the metric types.
+"""
+
+
 from .boolean import BooleanMetricType
 from .counter import CounterMetricType
 from .datetime import DatetimeMetricType
 from .event import EventMetricType, RecordedEventData
 from .experiment import RecordedExperimentData
-from .labeled import LabeledCounterMetricType
+from .labeled import (
+    LabeledBooleanMetricType,
+    LabeledCounterMetricType,
+    LabeledStringMetricType,
+)
 from .lifetime import Lifetime
+from .memory_distribution import MemoryDistributionMetricType
+from .memoryunit import MemoryUnit
 from .ping import PingType
 from .string import StringMetricType
+from .string_list import StringListMetricType
 from .timespan import TimespanMetricType
 from .timeunit import TimeUnit
+from .timing_distribution import TimingDistributionMetricType
 from .uuid import UuidMetricType
 
 
@@ -22,13 +35,19 @@ __all__ = [
     "CounterMetricType",
     "DatetimeMetricType",
     "EventMetricType",
+    "LabeledBooleanMetricType",
     "LabeledCounterMetricType",
+    "LabeledStringMetricType",
     "Lifetime",
+    "MemoryDistributionMetricType",
+    "MemoryUnit",
     "PingType",
     "RecordedEventData",
     "RecordedExperimentData",
     "StringMetricType",
+    "StringListMetricType",
     "TimespanMetricType",
     "TimeUnit",
+    "TimingDistributionMetricType",
     "UuidMetricType",
 ]