+++ gradle-plugin/src/main/groovy/mozilla/telemetry/glean-gradle-plugin/GleanGradlePlugin.groovy
@@ -34,7 +34,7 @@ class GleanMetricsYamlTransform extends ArtifactTransform {
 @SuppressWarnings("GrPackage")
 class GleanPlugin implements Plugin<Project> {
     // The version of glean_parser to install from PyPI.
-    private String GLEAN_PARSER_VERSION = "1.18.2"
+    private String GLEAN_PARSER_VERSION = "1.19.0"
     // The version of Miniconda is explicitly specified.
     // Miniconda3-4.5.12 is known to not work on Windows.
     private String MINICONDA_VERSION = "4.5.11"
@@ -330,10 +330,6 @@ subprocess.check_call([
         project.preBuild.dependsOn(createBuildDir)
         project.preBuild.finalizedBy("build_envs")
 
-        project.clean {
-            delete condaBootstrapDir
-        }
-
         return condaDir
     }
 
@@ -373,7 +369,7 @@ subprocess.check_call([
     }
 
     void apply(Project project) {
-        project.ext.glean_version = "25.1.0"
+        project.ext.glean_version = "26.0.0"
 
         File condaDir = setupPythonEnvironmentTasks(project)
         project.ext.set("gleanCondaDir", condaDir)